AG-8876 Move types into contextMenuTypes.ts
