+++ packages/ag-charts-community/src/chart/interaction/contextMenuRegistry.ts
@@ -1,111 +1,8 @@
-import type { RequireOptional } from 'ag-charts-core';
-import type {
-    AgContextMenuItem,
-    AgContextMenuItemAlways,
-    AgContextMenuItemLegendItem,
-    AgContextMenuItemLiteral,
-    AgContextMenuItemSeriesArea,
-    AgContextMenuItemShowOn,
-} from 'ag-charts-types';
+import type { AgContextMenuItemLiteral, AgContextMenuItemShowOn } from 'ag-charts-types';
 
 import { Listeners } from '../../util/listeners';
 import type { ContextMenuCallback, ContextMenuEvent, ContextShowOnMap } from './contextMenuTypes';
-
-type BuiltinItemListKeys = 'defaults';
-type BuiltinHideableKeys =
-    | 'zoom-to-cursor'
-    | 'pan-to-cursor'
-    | 'reset-zoom'
-    | 'toggle-series-visibility'
-    | 'toggle-other-series';
-
-type ContextMenuBuiltinItemsRules = {
-    readonly [K in Exclude<AgContextMenuItem, object | BuiltinItemListKeys>]: RequireOptional<AgContextMenuItem>;
-};
-type ContextMenuBuiltinItemListsRules = {
-    readonly [K in BuiltinItemListKeys]: readonly (keyof ContextMenuBuiltinItemsRules)[];
-};
-
-class ContextMenuBuiltinItems implements ContextMenuBuiltinItemsRules {
-    readonly download: RequireOptional<AgContextMenuItemAlways> = {
-        type: 'action',
-        showOn: 'always',
-        label: 'contextMenuDownload',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-    readonly 'zoom-to-cursor': RequireOptional<AgContextMenuItemSeriesArea> = {
-        type: 'action',
-        showOn: 'series-area',
-        label: 'contextMenuZoomToCursor',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-    readonly 'pan-to-cursor': RequireOptional<AgContextMenuItemSeriesArea> = {
-        type: 'action',
-        showOn: 'series-area',
-        label: 'contextMenuPanToCursor',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-    readonly 'reset-zoom': RequireOptional<AgContextMenuItemSeriesArea> = {
-        type: 'action',
-        showOn: 'series-area',
-        label: 'contextMenuResetZoom',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-    readonly 'toggle-series-visibility': RequireOptional<AgContextMenuItemLegendItem> = {
-        type: 'action',
-        showOn: 'legend-item',
-        label: 'contextMenuToggleSeriesVisibility',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-    readonly 'toggle-other-series': RequireOptional<AgContextMenuItemLegendItem> = {
-        type: 'action',
-        showOn: 'legend-item',
-        label: 'contextMenuToggleOtherSeries',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-    readonly 'separator': RequireOptional<AgContextMenuItemAlways> = {
-        type: 'separator',
-        showOn: 'always',
-        label: 'separator',
-        enable: true,
-        iconUrl: undefined,
-        action: undefined,
-        items: undefined,
-    };
-}
-
-class ContextMenuBuiltinItemLists implements ContextMenuBuiltinItemListsRules {
-    readonly defaults: readonly (keyof ContextMenuBuiltinItemsRules)[] = [
-        'download',
-        'zoom-to-cursor',
-        'pan-to-cursor',
-        'toggle-series-visibility',
-        'toggle-other-series',
-    ];
-}
-
-class ContextMenuBuiltins {
-    readonly items = new ContextMenuBuiltinItems();
-    readonly lists = new ContextMenuBuiltinItemLists();
-}
+import { ContextMenuBuiltins } from './contextMenuTypes';
 
 export class ContextMenuRegistry {
     public readonly builtins = new ContextMenuBuiltins();
@@ -158,7 +55,7 @@ export class ContextMenuRegistry {
         return !this.hiddenActions.has(id);
     }
 
-    public setVisible(id: BuiltinHideableKeys, visible: boolean) {
+    public setVisible(id: AgContextMenuItemLiteral, visible: boolean) {
         if (visible) {
             this.hiddenActions.delete(id);
         } else {
